Improve output
git-svn-id: http://root.cern.ch/svn/roottest/trunk@2023 7c735d0d-823a-0410-acfd-cd159556d4bf

+++ scripts/handleError.sh
@@ -34,7 +34,7 @@ if [ "x$logfile" != "x" ] ; then
      echo "'root.exe -b -l -q $testname' exited with error code: $result" >> $logfile
   fi
 fi
-if [ "x$toremove" != "x" ] ; then 
+if [ "x$toremove" != "x" -a -e $toremove ] ; then 
   echo handleError.sh: '*** Deleting file' $toremove
   rm $toremove
 fi
